Navigation added
The primary navigation scaffolding has been added. A ul within a div
with a class of 'navigation' will function as a main menu. Variables
added to tweak the menu settings and colors. Default is left-aligned but
'centered' and 'right' classes can be added to the 'navigation' div to
re-position the menu items.